More Updates
Made some changes to the GUI_window. Notably, I’ve added buttons to the
right of every input line. This will come in handy for when we add
gates and have to select which we want to use as inputs.

I also created TT_window, which is a separate window we can use for
displaying the truth table(s). I have not actually done any work with
this.