Major update - Gates and GUI
Gate class is a go save for istream input. I had to give the GUI a bit
of an overwrite to get things to work, but you will find the
application is now capable of adding a gate after picking 1 to 2 of the
existing lines. Still working on the code but I am not aware of
important bugs at the moment.